+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: support `zeebe:calledDecision binding property` ([#165](https://github.com/camunda/element-templates-json-schema/pull/165)
+
 * `FEAT`: support `zeebe:formDefinition` binding ([#169](https://github.com/camunda/element-templates-json-schema/pull/169)
 
 ## 0.23.0